Split NullableNumericAssertionSpecs
